feat(persistent): migrate LadybugDB engine lbug 0.15.4 -> 0.17.1 (lossless v40->v41 store upgrade) (#100) (#105)
Bumps the persistent GraphStore backend's pinned engine from `lbug = "=0.15.4"`
to `lbug = "=0.17.1"`. The 0.15.4 -> 0.17.1 Rust binding delta is purely
additive, so no source changes are needed in graph/lbug_store — only the pin and
a recompile (verified: `cargo build --features persistent` is clean).

No data loss (verified empirically, CI-blocking):
- lbug 0.17.1's `StorageVersionInfo::canReadStorageVersion(40)` returns true and
  the catalog deserializers carry explicit `else { upgradeLegacyStorageFormat() }`
  branches for storage version < 41, so a v40 store (the live ~/.simard/cognitive
  format) opens IN PLACE under 0.17.1 — no dump/reload, no quarantine, no
  rebuild-from-empty. The first checkpoint upgrades the on-disk store to v41
  (forward-only).
- Adds a committed v40 fixture (captured from a real 0.15.4 build) and a
  regression test proving it opens `Clean` (never `RebuiltAfterCorruption`) with
  every node/edge intact, then upgrades to v41 on checkpoint.
- A dedicated **blocking** CI step (`graph::lbug_store::tests::v40`) now gates
  merges on the no-data-loss proof. It links the prebuilt 0.17.x engine archive
  (fast, deterministic), so blocking adds no flakiness. The broader persistent
  suite (probabilistic CSR churn) stays continue-on-error.

CSR SIGSEGV (#100) not regressed: the version-independent soft-delete tombstone
keeps the store off the crashing CSR-delete path on every engine version;
`reproduces_issue_100_csr_delete_corruption` stays green on 0.17.1 (3/3 reruns).

Docs: new docs/lbug_0_17_upgrade.md (full migration reference + live-store
runbook + Simard coordination), linked from README and safe_node_deletion.md.

Incidental: fixes three pre-existing test-only clippy nits surfaced by the
`--all-targets --features persistent` gate (unnecessary cast, module_inception,
map_or->is_none_or) so that gate is genuinely green.

amplihack-memory-lib stays on the unpatched crates.io `lbug` (remains
publishable); the optional C++ CSR defense-in-depth patch is applied only by
Simard via [patch.crates-io] (PR #2, separate checkout).

+++ rust/amplihack-memory/tests/concurrent.rs
@@ -494,7 +494,7 @@ fn test_concurrent_sqlite_backend() {
     let b = backend.lock().unwrap();
     let stats = b.get_statistics().unwrap();
     assert_eq!(
-        stats.total_experiences as usize,
+        stats.total_experiences,
         threads * ops,
         "expected {} total experiences, got {}",
         threads * ops,
